lottery 2

+++ proc.c
@@ -323,7 +323,7 @@ static
 unsigned long
 lcg_rand(unsigned long a)
 {
-  unsigned long b=279470273,c=4294967291;
+  int b=279470273 , c=4294967291;
   return(a*b)%c;
 }
 int lotteryTotal(void){
